changed avatar EP calls
