+++ src/api/repositories/profileRepository/profileRepository.api.endpoints.ts
@@ -4,6 +4,7 @@ export const userRepositoryEndpoints = () => {
   const BASE_URL = `${appVariables.BACKEND_ENDPOINT_URL}/profile`;
 
   return {
-    edit: `${BASE_URL}/edit`
+    edit: `${BASE_URL}/edit`,
+    avatarUpload: `${BASE_URL}/avatar/upload`
   };
 };
