+++ src/api/repositories/profileRepository/profileRepository.api.types.ts
@@ -10,3 +10,13 @@ export interface UpdateUserRequest {
 export interface UpdateUserResponse {
   userOnboarded?: boolean;
 }
+
+/** UPLOAD AVATAR **/
+
+export interface UploadAvatarRequest {
+  avatar: File;
+}
+
+export interface UploadAvatarResponse {
+  hash: string;
+}
