+++ src/api/repositories/userRepository/userRepository.api.endpoints.ts
@@ -6,7 +6,6 @@ export const userRepositoryEndpoints = () => {
   return {
     me: `${BASE_URL}/me`,
     check: `${BASE_URL}/check`,
-    avatarUpload: `${BASE_URL}/avatar/upload`,
     inviteToSpace: `${BASE_URL}/invite`,
     search: `${BASE_URL}/search`,
     profile: `${BASE_URL}/profile`,
